FEATURE: allow \external functions
it is now possible to declare functions as external such that the names
of these functions are not masked when passing to mathematica. That way
Sqrt function can be declared as \external R Sqrt(R); for instance.
BEWARE: this not yet used for functions inside of programs so there so
there might be issues with e.g. differential equation systems containing
these symbols. It should be fine if they appear in assignments and tests
though.