FEATURE: added DA rule
note that you cannot use #dlvar in phi but you have to instead use n
there. as this introduces quantified terms I added the rules
all_singleton that allows to remove a universal quantifier if it is of
the form \forall u; (u = q -> phi) and u does not occur in q.