Initial checkin of aircraft elevation control example
The example is based on the MatLab Simulink model sf_aircraft. Currently the formalization captures sensor readout and validation for the left outer position sensors. The header files contain macro definitions used for a parallel execution of the two controllers next to the physics. A simpler example of this package can be found in parallel_increase.key.