BUGFIX: fix soundness w.r.t. local QE and redlog
Release 3.6.2:
Local quantifier elimination must not use the universal closure since it
does not yield an equivalent transformation anymore. Universal closure
before QE only works for global QE since if the closure is a tautology
so is the original formula. However, if the closure is equivalent to
false it does not mean that the original formula is unsatisfyable.